portal/gpio/requirements.txt
RPi.GPIO==0.5.4
wsgiref==0.1.2
